Added Price property data validation to not accept negative values. Moved length validation to model. Both UpdateItem and DeleteItem now check if item exists before taking action.

+++ BackEndPracticalTest/FactoryAPI/Controllers/ItemsController.cs
@@ -23,15 +23,10 @@ namespace FactoryAPI.Controllers
         [ProducesResponseType(500)]
         public async Task<ActionResult> CreateItem([FromBody] Item item)
         {
-            //check item validation
-            if (item.Name.Length > 50)
+            if(item.Price < 0)
             {
-                ModelState.AddModelError("Error", "Length of 'Name' property should not exceed 50 characters");
-            }
-
-            if (item.Description.Length > 100)
-            {
-                ModelState.AddModelError("Error", "Length of 'Description' property should not exceed 100 characters");
+                ModelState.AddModelError("Error", "You should be possitive in your life, especially when setting the price!");
+                return StatusCode(400, ModelState);
             }
 
             if (!ModelState.IsValid)
@@ -46,7 +41,7 @@ namespace FactoryAPI.Controllers
                 return StatusCode(500, ModelState);
             }
 
-            return Ok("Successfully Created!");
+            return Ok(item.ID);
         }
 
         // Retrieve a single item
@@ -80,19 +75,23 @@ namespace FactoryAPI.Controllers
         [ProducesResponseType(500)]
         public async Task<ActionResult> UpdateItem([FromBody] Item item)
         {
-            if (item.Name.Length > 50)
-            {
-                ModelState.AddModelError("Error", "Length of 'Name' property should not exceed 50 characters");
-            }
-
-            if (item.Description.Length > 100)
-            {
-                ModelState.AddModelError("Error", "Length of 'Description' property should not exceed 100 characters");
-            }
-
             //check item validation
+            if (item.Price < 0)
+            {
+                ModelState.AddModelError("Error", "You should be possitive in your life, especially when setting the price!");
+                return StatusCode(400, ModelState);
+            }
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            var dbItem = await _itemRepository.GetItemById(item.ID);
+
+            if (dbItem == null)
+            {
+                ModelState.AddModelError("Error", "Couldn't find the item");
+                return NotFound(ModelState);
+            }
 
             var result = await _itemRepository.Update(item);
 
@@ -102,7 +101,7 @@ namespace FactoryAPI.Controllers
                 return StatusCode(500, ModelState);
             }
 
-            return Ok("Successfully Updated!");
+            return Ok(item.ID);
         }
 
         // Delete an item
@@ -114,6 +113,14 @@ namespace FactoryAPI.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            var dbItem = await _itemRepository.GetItemById(id);
+
+            if(dbItem == null)
+            {
+                ModelState.AddModelError("Error", "Couldn't find the item");
+                return NotFound(ModelState);
+            }
 
             var result = await _itemRepository.Delete(id);
